requirements.txt
pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
pip install git+https://github.com/openai/whisper.git
ffmpeg-python
opencv-python
paddleocr
ultralytics
langchain
openai
python-dotenv
rich
